refactor: centralize mock factories
