+++ tests/common/mockFactories/index.js
@@ -2,3 +2,9 @@ export * from './coreServices.js';
 export * from './loaders.js';
 export * from './entities.js';
 export * from './container.js';
+
+// Explicit convenience exports
+export {
+  createMockPathResolver,
+  createMockDataFetcher,
+} from './coreServices.js';
